qidian/swipe/requirements.txt
Pillow
opencv-python
pyyaml
numpy
configargparse